Updated entites

+++ praksa-dao/src/main/java/ba/com/zira/praksa/dao/model/ReviewGradeEntity.java
@@ -39,7 +39,7 @@ public class ReviewGradeEntity implements Serializable {
     @Column(name = "formula_id")
     private Long formulaId;
 
-    private Long grade;
+    private Double grade;
 
     private String type;
